+++ Standalone/FrequencyGame.cpp
@@ -285,7 +285,7 @@ Effects frequency_game_update(FrequencyGame_State *state, Event event)
             state->current_file_idx = juce::Random::getSystemRandom().nextInt((int)state->files.size());
             effects.player = Effect_Player {
                 .commands = { 
-                    { .type = Audio_Command_Load, .value_file = state->files[state->current_file_idx].file },
+                    { .type = Audio_Command_Load, .value_file = state->files[static_cast<size_t>(state->current_file_idx)].file },
                     { .type = Audio_Command_Play },
                 }
             };
